+++ flink/src/main/java/com/jie/flink/function/WaterSensorMapFunction.java
@@ -12,6 +12,6 @@ public class WaterSensorMapFunction implements MapFunction<String, WaterSensor> 
     @Override
     public WaterSensor map(String line) throws Exception {
         String[] split = line.split(",");
-        return new WaterSensor(split[0],Integer.parseInt(split[1]),Integer.parseInt(split[2]));
+        return new WaterSensor(split[0],Long.parseLong(split[1]),Integer.parseInt(split[2]));
     }
 }
